requirements.txt
deap==1.4.2
hypothesis==6.142.3
matplotlib==3.10.7
networkx==3.3
numpy==2.3.4
pandas==2.3.3
pipreqs==0.5.0
pyswarms==1.3.0
pytest==7.4.4
pytest.egg==info
PyYAML==6.0.1
PyYAML==6.0.3
scipy==1.16.2
tqdm==4.66.5
